pyproject.toml
[tool.poetry]
name = "tap-domo"
version = "0.1.0"
description = "A Singer Tap by Mashey"
authors = ["Ryan Allen <rcallen89@gmail.com>"]
packages = [
    { include = "tap_domo" }
]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
singer-python = "^5.9.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
coverage = "^5.4"
pylint = "^2.6.0"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
coveralls = "^3.0.0"
pytest-mock = "^3.5.1"
pytest-vcr = "^1.0.2"
python-dotenv = "^0.15.0"
vcrpy = "^4.1.1"

[tool.poetry.scripts]
tap-domo = 'tap_domo:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"